refector:去除部分调试日志

+++ ios-ui/HelloTako/HelloTako/detailView/GameDetailViewController.m
@@ -87,7 +87,7 @@
 -(IBAction)  segmentAction:(UISegmentedControl *)Seg{
     NSInteger Index = Seg.selectedSegmentIndex;
     
-    NSLog(@"Index %i", Index);
+    NSLog(@"Index %li", (long)Index);
     
     switch (Index) {
         case 0:
